requirements.txt
psycopg2-binary
requests
python-dotenv
beautifulsoup4
streamlit
cohere
flask==3.0.0
pyyaml
PyPDF2
mistralai
numpy